requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# CLI interface
click>=8.0.0
rich>=13.0.0  # For colored output and better formatting

# Configuration management
python-dotenv>=1.0.0

# Optional: LangSmith for evaluation (can be added later)
# langsmith>=0.1.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
